hubungin transaksi ke role lain

+++ database/migrations/2025_11_09_163742_create_review_table.php
@@ -11,30 +11,33 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('reviews', function (Blueprint $table) {
-            $table->bigIncrements('review_id');
-            $table->unsignedBigInteger('user_id');
-            $table->unsignedBigInteger('product_id');
-            $table->integer('rating');
-            $table->text('comment')->nullable();
-            $table->dateTime('date')->nullable();
-            $table->timestamps();
-            $table->index('user_id');
-            $table->index('product_id');
-            
-            $table->foreign('user_id')
-                ->references('user_id')
-                ->on('users')
-                ->onDelete('cascade')
-                ->onUpdate('cascade');
-                
-            $table->foreign('product_id')
-                ->references('product_id')
-                ->on('products')
-                ->onDelete('cascade')
-                ->onUpdate('cascade');
-            
-        });
+        if (!Schema::hasTable('reviews')) {
+
+            Schema::create('reviews', function (Blueprint $table) {
+                $table->bigIncrements('review_id');
+                $table->unsignedBigInteger('user_id');
+                $table->unsignedBigInteger('product_id');
+                $table->integer('rating');
+                $table->text('comment')->nullable();
+                $table->dateTime('date')->nullable();
+                $table->timestamps();
+
+                $table->index('user_id');
+                $table->index('product_id');
+
+                $table->foreign('user_id')
+                    ->references('user_id')
+                    ->on('users')
+                    ->onDelete('cascade')
+                    ->onUpdate('cascade');
+
+                $table->foreign('product_id')
+                    ->references('product_id')
+                    ->on('products')
+                    ->onDelete('cascade')
+                    ->onUpdate('cascade');
+            });
+        }
     }
 
     /**
@@ -42,10 +45,14 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('reviews', function (Blueprint $table) {
-            $table->dropForeign(['user_id']);
-            $table->dropForeign(['product_id']);
-        });
-        Schema::dropIfExists('reviews');
+        if (Schema::hasTable('reviews')) {
+            Schema::table('reviews', function (Blueprint $table) {
+                // Hapus Foreign Key dulu (praktik yang aman)
+                $table->dropForeign(['user_id']);
+                $table->dropForeign(['product_id']);
+            });
+
+            Schema::dropIfExists('reviews');
+        }
     }
 };
